Fixing new google analytics card design to work properly with atomic
